pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "banking-medallion-lakehouse"
version = "0.1.0"
description = "End-to-end medallion lakehouse scaffold with Delta Lake, Airflow, Spark, Trino, and dbt."
requires-python = ">=3.11"
dependencies = [
  "boto3>=1.35.0",
  "delta-spark>=3.2.0",
  "fastavro>=1.9.0",
  "Faker>=27.0.0",
  "kafka-python>=2.0.2",
  "pendulum>=3.0.0",
  "psycopg[binary]>=3.2.0",
  "pyspark>=3.5.2",
  "PyYAML>=6.0.2",
  "requests>=2.32.0",
  "tenacity>=9.0.0",
  "trino>=0.329.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11.0",
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = [
  "ingestion*",
  "lakehouse*",
  "quality*",
  "optimization*",
  "observability*",
  "orchestration*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

